requirements.txt
mss==9.0.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
six==1.16.0
